Redistribuição de funções em diversos arquivos
19/05/2023
Arquivos:
- Redistribuição de todas as funções dos arquivos: menu.c, cadastro.c e interface.c, para arquivos separados e com apenas uma função.
- Criação de vários arquivos, em destaque, a alternância da função main, antes em menu.c para o arquivo main.c.
- Melhora na codificação do .gitignore, para incluir todos os arquivos ".o" e ".exe".
