+++ .gitignore
@@ -1,16 +1,4 @@
-menu.exe
-menu.o
-cadastro.o
-interface.o
-testeCadastro.exe
-SistemaControleEmpresarial.exe
 Makefile.win
-main.o
-ProjetoTesteInterface.exe
-testeInterface.exe
-programa.exe
-global_var.o
-Projeto1.exe
 Projeto1.layout
-testeInicialConexao.exe
-testeInicialConexao.o
+*.o
+*.exe
