+++ SistemaControleEmpresarial.layout
@@ -1,68 +1,38 @@
-[Editor_0]
-CursorCol=12
-CursorRow=323
-TopLine=283
-LeftChar=1
 [Editors]
-Order=4,1,6,9,0,2,3,8,7
-Focused=1
-[Editor_1]
-CursorCol=25
-CursorRow=8
+Order=28
+Focused=28
+[Editor_28]
+CursorCol=15
+CursorRow=20
 TopLine=1
 LeftChar=1
-[Editor_2]
-CursorCol=27
-CursorRow=164
-TopLine=142
+[Editor_14]
+CursorCol=1
+CursorRow=1
+TopLine=11
 LeftChar=1
-[Editor_3]
-CursorCol=7
-CursorRow=140
-TopLine=27
-LeftChar=1
-[Editor_4]
-CursorCol=17
-CursorRow=13
-TopLine=1
-LeftChar=1
-[Editor_5]
-CursorCol=4
-CursorRow=63
-TopLine=1
-LeftChar=1
-[Editor_6]
-CursorCol=7
-CursorRow=16
-TopLine=1
-LeftChar=1
-[Editor_7]
-CursorCol=33
-CursorRow=11
+[Editor_16]
+CursorCol=1
+CursorRow=1
 TopLine=1
 LeftChar=1
 [Editor_8]
-CursorCol=23
-CursorRow=84
-TopLine=76
+CursorCol=1
+CursorRow=1
+TopLine=34
 LeftChar=1
 [Editor_9]
 CursorCol=1
-CursorRow=8
-TopLine=1
+CursorRow=1
+TopLine=20
 LeftChar=1
-[Editor_10]
-CursorCol=30
-CursorRow=5
-TopLine=1
-LeftChar=1
-[Editor_11]
-CursorCol=43
-CursorRow=17
+[Editor_15]
+CursorCol=1
+CursorRow=1
 TopLine=1
 LeftChar=1
 [Editor_12]
-CursorCol=7
-CursorRow=23
+CursorCol=1
+CursorRow=1
 TopLine=1
 LeftChar=1
