+++ cadastro/cadastro_funcoes/cadastro_posicaoDisponivel.c
@@ -1,0 +1,15 @@
+#include "../../global/global.h"
+#include "../cadastro.h"
+
+int posicaoDisponivel(int *vetorReferencia) {
+	int i;
+	
+	for(i = 0; i < MAX_VETOR; i++) {
+		if(vetorReferencia[i] == 0) {
+			vetorReferencia[i] = 1;
+			break;
+		}
+	}
+	
+	return i;
+}
