+++ cadastro/cadastro_funcoes/cadastro_validarPalavraChave.c
@@ -1,0 +1,14 @@
+#include "../../global/global.h"
+#include "../cadastro.h"
+
+bool validarPalavraChave(char string[], char stringChave[]) {
+	char stringHolder[MAX_STRING];
+	int i;
+	
+	for(i = 0; i < strlen(stringChave); i++) {
+		stringHolder[i] = toupper(string[i]);
+	}
+	
+	if(strncmp(stringHolder, stringChave, strlen(stringChave)) == 0) return true;
+	else return false;
+}
