+++ cadastro/cadastro_funcoes/cadastro_verificarDadosVetorDisponivel.c
@@ -1,0 +1,10 @@
+#include "../../global/global.h"
+#include "../cadastro.h"
+
+void verificarDadosVetorDisponivel(int *vetorReferencia, int posicaoConsultada) {
+	posicaoConsultada -= 1;
+	
+	if(vetorReferencia[posicaoConsultada] == 0) {
+		erro = 4;
+	}
+}
