+++ interface/interface_funcoes/interface_exibirInterfaceInteracao.c
@@ -1,0 +1,17 @@
+#include "../../global/global.h"
+#include "../interface.h"
+
+void exibirInterfaceInteracao(char string[]) {
+	int i;
+	int tamanhoString = strlen(string); 
+	
+	for(i = 0; i < strlen(string) + 4; i++) {
+		printf(".");
+	}
+		printf("\n| %s |\n", string);
+	for(i = 0; i < strlen(string) + 4; i++) {
+		printf(".");
+	}
+	
+	printf("\n\n");
+}
