+++ interface/interface_funcoes/interface_exibirLogotipo.c
@@ -1,0 +1,12 @@
+#include "../../global/global.h"
+#include "../interface.h"
+
+void exibirLogotipo() {
+	int i;
+	
+	for(i = 0; i < 6;i++) {
+		printf("%s\n", displayRefLogoTipo[i]);
+	}
+	printf("\n");
+
+}
