+++ interface/interface_var/interface_var.c
@@ -1,0 +1,106 @@
+#include "interface_var.h"
+
+Display displayRefFormularioCadastroCliente[13] = {
+	"+----------------------------------------------------------------------------------------------------+",
+	"| Num Cadastro:     | [1] Nome:                                                                      |", //NumCadastro: 16 // Nome: 32 -> 100
+	"+----------------------------------------------------------------------------------------------------+", 
+	"| [2] CPF:                                   | [3] Data Nasc.:                                       |", //CPF: 11 -> 44 // Data Nasc.: 63 -> 100
+	"+----------------------------------------------------------------------------------------------------+",
+	"| [4] Telefone:                              | [5] E-mail:                                           |", //Telefone: 16 -> 44 // E-mail: 59 -> 100
+	"+----------------------------------------------------------------------------------------------------+",
+	"| [6] Endereço:                                                                | [7] Número:         |", //Endereço: 16 -> 78 // Número: 93 -> 100
+	"+----------------------------------------------------------------------------------------------------+",
+	"| [8] Complemento:                           | [9] Bairro:                                           |", //Complemento: 19 -> 44 // Bairro: 59 -> 100
+	"+----------------------------------------------------------------------------------------------------+",
+	"| [10] CEP:                 | [11] Cidade:                                     | [12] Estado:        |", //CEP: 12 -> 26 // Cidade: 43 -> 90// Estado: 94 -> 100 
+	"+----------------------------------------------------------------------------------------------------+",
+};
+
+Display displayRefFormularioCadastroFuncionario[15] = {
+	"+----------------------------------------------------------------------------------------------------+",
+	"| Num Cadastro:     | [1] Nome:                                                | [2] Status:         |", //NumCadastro: 16-> 19 // Nome: 32 ->  // Status: 93
+	"+----------------------------------------------------------------------------------------------------+",
+	"| [3] CPF:                        | [4] Data de Nasc.:            | [5] Estado Civil:                |", //CPF: 11 -> 32 // Data de Nasc.: 55 -> 64 // Estado Civil: 86 -> 100
+	"+----------------------------------------------------------------------------------------------------+",
+	"| [6] Data de Admissão:                       | [7] Salário Base: R$                                 |", //Data de Admissão: 24 -> 44 // Salário Base: 68 -> 100
+	"+----------------------------------------------------------------------------------------------------+",
+	"| [8] Telefone:                               | [9] E-mail:                                          |", //Telefone: 16 -> 44 // E-mail: 60 -> 100
+	"+----------------------------------------------------------------------------------------------------+",
+	"| [10] Endereço:                                                               | [11] Número:        |", //Endereço: 17 -> 78 // Número: 94 -> 100
+	"+----------------------------------------------------------------------------------------------------+",
+	"| [12] Complemento:                           | [13] Bairro:                                         |", //Complemento: 20 -> 44 // Bairro: 61 -> 100
+	"+----------------------------------------------------------------------------------------------------+",
+	"| [14] CEP:                 | [15] Cidade:                                     | [16] Estado:        |", //CEP: 12 -> 26 // Cidade: 43 -> // Estado: 94 -> 100 
+	"+----------------------------------------------------------------------------------------------------+",
+};
+
+Display displayRefFormularioCadastroFornecedor[11] = {
+	"+----------------------------------------------------------------------------------------------------+",
+	"| Num Cadastro:     | [1] Nome Fantasia:                                                             |", //NumCadastro: 16-> 19 // Nome Fantasia: 41 -> 100
+	"+----------------------------------------------------------------------------------------------------+",
+	"| [2] CNPJ:                    | [3] Telefone:                  | [4] E-mail:                        |", //CNPJ: 12 -> 26 // Telefone: 47 -> 59 // E-mail: 78 -> 100
+	"+----------------------------------------------------------------------------------------------------+",
+	"| [5] Endereço:                                                                | [6] Número:         |", //Endereço: 16 -> 78 // Número: 93 -> 100
+	"+----------------------------------------------------------------------------------------------------+",
+	"| [7] Complemento:                           | [8] Bairro:                                           |", //Complemento: 19 -> 44 // Bairro: 59 -> 100
+	"+----------------------------------------------------------------------------------------------------+",
+	"| [8] CEP:                  | [9] Cidade:                                      | [10] Estado:        |", //CEP: 11-> 26 // Cidade: 42 -> 89// Estado: 94 -> 100 
+	"+----------------------------------------------------------------------------------------------------+",
+};
+
+Display displayRefTitulo[3] = {
+	"======================================================================================================",
+	"||                                                                                                  ||",
+	"======================================================================================================",
+};
+
+Display displayRefCabecalhoConsultaClientes[MAX_CABECALHO] = {
+	"+----------------------------------------------------------------------------------------------------+",
+	"| Nº Cadastro |                Nome                   |       CPF       |          Telefone          |",
+	"+----------------------------------------------------------------------------------------------------+",
+};
+
+Display displayRefCabecalhoConsultaFuncionarios[MAX_CABECALHO] = {
+	"+----------------------------------------------------------------------------------------------------+",
+	"|  Status  | Nº Cadastro |                Nome                   |       CPF       |     Telefone    |",
+	"+----------------------------------------------------------------------------------------------------+",
+};
+
+Display displayRefCabecalhoConsultaFornecedores[MAX_CABECALHO] = {
+	"+----------------------------------------------------------------------------------------------------+",
+	"| Nº Cadastro |             Nome Fantasia             |         CNPJ          |       Telefone       |",
+	"+----------------------------------------------------------------------------------------------------+",
+};
+
+Display displayRefConsulta_Cliente[2] = {
+	"|             |                                       |                 |                            |", //7 //16 -> 50
+	"+----------------------------------------------------------------------------------------------------+",
+};
+
+Display displayRefConsulta_Funcionario[2] = {
+	"|          |             |                                       |                 |                 |", //7 //13 //26 -> 63 // 67 // 86
+	"+----------------------------------------------------------------------------------------------------+",
+};
+
+Display displayRefConsulta_Fornecedor[2] = {
+	"|             |                                       |                       |                      |", //7 //16 -> 50
+	"+----------------------------------------------------------------------------------------------------+",
+};
+
+Display displayRefConsulta_SemDados[2] = {
+	"|                                     Não há nenhum dado ainda.                                      |", 
+	"+----------------------------------------------------------------------------------------------------+",
+};
+
+Display displayRefLogoTipo[6] = {
+	"                  _____           _       ______                     _               ",
+	"                 |  __ \\         | |     |  ____|                   | |             ",
+	"                 | |__) |   ___  | |_    | |__      ___    ___    __| |   ___   _ __ ",
+	"                 |  ___/   / _ \\ | __|   |  __|    / _ \\  / _ \\  / _` |  / _ \\ | '__|",
+	"                 | |      |  __/ | |_    | |      |  __/ |  __/ | (_| | |  __/ | |   ",
+	"                 |_|       \\___|  \\__|   |_|       \\___|  \\___|  \\__,_|  \\___| |_|",
+	
+};
+
+int tamanhoDisplay;
+int contadorExistentes = 0;
